requirements.txt
Flask
numpy
opencv-python
pillow
PyYAML
ultralytics